modify 'sms' demo-module according to new lib.
